Update patch set 1
Patch Set 1:

Giacomo Longo	:  Do you have proof that it does? 

My reservations with this patch is that a) only a small subset of people need/want this  b) usability is a bit broken 

Be nice to toggle/hide via a setting, then again only a hand full of people are going to be using sip calling on their N7s (which i guess could apply to wifi tilapias and other devices as well)

One thing I really don't want to see is device specific changes for something that isn't device specific.  This should be something to be reviewed/looked at for any 3G tablet (or all tablets?)  I hate seeing device specific features "fragmenting" cyanogenmod (ok that was a bad word, but you get the idea.

Patch-set: 1