Update patch set 1
Patch Set 1:

@Janson Kang	There is a difference between nerd knob and features.   Just cause you can doesn't mean it's a useful feature.  This doesn't mean that I don't think you could come up with some reasoning, it's just not that required.  If there are more who request SMS/MMS maybe we can look into it but at this point it's not really a feature.  It's a nerd knob. 

Nerd knobs are fine and nerds can compile their own just like you did to get those in.  No clutter in functionality.   No, less than apt, users wondering why their mms isn't working on a plan that doesn't support it, etc.

The one way I may look into this feature is if we can freeze/hide it by default and if it applies to every GSM tablet on Cyanogenmod (and signed off by those who control that).  It is also important that base features like this be common across the project.  Rule of thumb usually is, if it isn't hardware specific it should be done at a higher layer.

TL;DR;  This isn't a daily use feature and doesn't enhance the tablet that much.  It's just a way to add minimal/zero use feature just cause you can.

Patch-set: 1